Consolidating the Asana struct and impl into the parent module file
